chore: Update to version 1.3.1 with bug fixes and performance improvements
- Fixed critical crash on the Settings page and improved error handling with user-friendly messages.
- Resolved loading issues in Image Library and Animated Library pages.
- Implemented async loading for enhanced performance and responsiveness.
- Added loading indicators and empty state messages for better user experience.
- Updated dependencies to the latest compatible versions.

**Status**: Production Ready | Free & Open Source

+++ AnimeQuoteWall.Core/Services/AnimationService.cs
@@ -162,7 +162,7 @@ public class AnimationService
                 image = new MagickImage(framePath);
 
                 // Set frame delay for GIF animation timing
-                image.AnimationDelay = delayInHundredths;
+                image.AnimationDelay = (uint)delayInHundredths;
 
                 // Set quality to reduce file size (90 is a good balance)
                 image.Quality = 90;
